AC-38:  Refactor data
