+++ data/api/src/main/java/com/jfmr/ac/test/data/api/rickandmorty/dto/character/entity/LocationResponse.kt
@@ -1,11 +1,11 @@
-package com.jfmr.ac.test.data.api.rickandmorty.character.entity
+package com.jfmr.ac.test.data.api.rickandmorty.dto.character.entity
 
 import com.google.gson.annotations.SerializedName
 import kotlinx.serialization.Serializable
 
 
 @Serializable
-data class OriginResponse(
+data class LocationResponse(
 
     @field:SerializedName("name")
     val name: String? = null,
