+++ data/api/src/main/java/com/jfmr/ac/test/data/api/rickandmorty/dto/character/entity/OriginResponse.kt
@@ -1,11 +1,11 @@
-package com.jfmr.ac.test.data.api.rickandmorty.character.entity
+package com.jfmr.ac.test.data.api.rickandmorty.dto.character.entity
 
 import com.google.gson.annotations.SerializedName
 import kotlinx.serialization.Serializable
 
 
 @Serializable
-data class LocationResponse(
+data class OriginResponse(
 
     @field:SerializedName("name")
     val name: String? = null,
