+++ data/api/src/main/java/com/jfmr/ac/test/data/api/rickandmorty/network/RickAndMortyApiService.kt
@@ -1,8 +1,8 @@
 package com.jfmr.ac.test.data.api.rickandmorty.network
 
-import com.jfmr.ac.test.data.api.rickandmorty.character.entity.CharacterResponse
-import com.jfmr.ac.test.data.api.rickandmorty.character.entity.CharactersResponse
-import com.jfmr.ac.test.data.api.rickandmorty.episode.entity.EpisodeResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.character.entity.CharacterResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.character.entity.CharactersResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.episode.entity.EpisodeResponse
 import com.jfmr.ac.test.data.api.rickandmorty.network.EndPoints.CHARACTER_DETAIL_URL
 import com.jfmr.ac.test.data.api.rickandmorty.network.EndPoints.CHARACTER_URL
 import com.jfmr.ac.test.data.api.rickandmorty.network.EndPoints.EPISODES_URL
