+++ data/cache/src/main/java/com/jfmr/ac/test/data/cache/datasource/LocalCharacterDataSourceImpl.kt
@@ -32,4 +32,8 @@ class LocalCharacterDataSourceImpl @Inject constructor(
 
     override suspend fun remoteKeysId(id: Long): RemoteKeys? =
         rickAndMortyDB.remoteKeysDao().remoteKeysId(id)
+
+    override suspend fun clearRemoteKeyByQuery(query: String) {
+        rickAndMortyDB.remoteKeysDao().clearRemoteKeys(query)
+    }
 }
