--- data/di/src/main/java/com/jfmr/ac/test/data/di/DBModule.kt
@@ -1,67 +1,0 @@
-package com.jfmr.ac.test.data.di
-
-
-import android.content.Context
-import com.jfmr.ac.test.data.cache.dao.character.CharacterDao
-import com.jfmr.ac.test.data.cache.dao.character.RemoteKeysDao
-import com.jfmr.ac.test.data.cache.dao.episode.EpisodeDao
-import com.jfmr.ac.test.data.cache.datasource.LocalCharacterDataSource
-import com.jfmr.ac.test.data.cache.datasource.LocalCharacterDataSourceImpl
-import com.jfmr.ac.test.data.cache.db.RickAndMortyDB
-import com.jfmr.ac.test.data.paging.RickAndMortyRemoteMediator
-import com.jfmr.ac.test.data.remote.character.datasource.CharacterRemoteDataSource
-import dagger.Binds
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module(includes = [DBModule.Declarations::class])
-@InstallIn(SingletonComponent::class)
-object DBModule {
-
-    @InstallIn(SingletonComponent::class)
-    @Module
-    interface Declarations {
-
-        @Singleton
-        @Binds
-        fun bindsLocalLoginDS(implementation: LocalCharacterDataSourceImpl): LocalCharacterDataSource
-
-    }
-
-    @Provides
-    @Singleton
-    fun provideDatabase(@ApplicationContext context: Context): RickAndMortyDB =
-        RickAndMortyDB.create(context)
-
-    @Provides
-    @Singleton
-    fun provideCharacterDao(database: RickAndMortyDB): CharacterDao {
-        return database.characterDao()
-    }
-
-    @Provides
-    @Singleton
-    fun provideRemoteKeyDao(database: RickAndMortyDB): RemoteKeysDao {
-        return database.remoteKeysDao()
-    }
-
-    @Provides
-    @Singleton
-    fun provideMediator(
-        localCharacterDataSource: LocalCharacterDataSource,
-        characterRemoteDataSource: CharacterRemoteDataSource,
-    ) = RickAndMortyRemoteMediator(
-        localCharacterDataSource = localCharacterDataSource,
-        characterRemoteDataSource = characterRemoteDataSource
-    )
-
-    @Provides
-    @Singleton
-    fun provideEpisodesDao(database: RickAndMortyDB): EpisodeDao {
-        return database.episodesDao()
-    }
-}
