+++ data/paging/src/main/java/com/jfmr/ac/test/data/paging/mapper/CharacterExtensions.kt
@@ -1,8 +1,8 @@
 package com.jfmr.ac.test.data.paging.mapper
 
-import com.jfmr.ac.test.data.api.rickandmorty.character.entity.CharacterResponse
-import com.jfmr.ac.test.data.api.rickandmorty.character.entity.LocationResponse
-import com.jfmr.ac.test.data.api.rickandmorty.character.entity.OriginResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.character.entity.CharacterResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.character.entity.LocationResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.character.entity.OriginResponse
 import com.jfmr.ac.test.data.cache.entities.character.LocalCharacter
 import com.jfmr.ac.test.data.cache.entities.character.LocalLocation
 import com.jfmr.ac.test.data.cache.entities.character.LocalOrigin
