+++ data/remote/src/main/java/com/jfmr/ac/test/data/remote/character/CharacterRemoteDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package com.jfmr.ac.test.data.remote.character
 
-import com.jfmr.ac.test.data.api.rickandmorty.character.entity.CharacterResponse
-import com.jfmr.ac.test.data.api.rickandmorty.character.entity.CharactersResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.character.entity.CharacterResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.character.entity.CharactersResponse
 import com.jfmr.ac.test.data.api.rickandmorty.network.RickAndMortyApiService
 import com.jfmr.ac.test.data.remote.character.datasource.CharacterRemoteDataSource
 import retrofit2.Response
