+++ data/remote/src/main/java/com/jfmr/ac/test/data/remote/character/datasource/CharacterRemoteDataSource.kt
@@ -1,7 +1,7 @@
 package com.jfmr.ac.test.data.remote.character.datasource
 
-import com.jfmr.ac.test.data.api.rickandmorty.character.entity.CharacterResponse
-import com.jfmr.ac.test.data.api.rickandmorty.character.entity.CharactersResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.character.entity.CharacterResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.character.entity.CharactersResponse
 import com.jfmr.ac.test.data.api.rickandmorty.network.RickAndMortyApiService
 import retrofit2.Response
