+++ data/remote/src/main/java/com/jfmr/ac/test/data/remote/di/DispatchersModule.kt
@@ -1,14 +1,21 @@
-package com.jfmr.ac.test.data.di
+package com.jfmr.ac.test.data.remote.di
 
-import com.jfmr.ac.test.data.remote.qualifier.DispatcherDefault
-import com.jfmr.ac.test.data.remote.qualifier.DispatcherIO
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
+import javax.inject.Qualifier
 import javax.inject.Singleton
+
+@Retention(AnnotationRetention.BINARY)
+@Qualifier
+annotation class DispatcherDefault
+
+@Retention(AnnotationRetention.BINARY)
+@Qualifier
+annotation class DispatcherIO
 
 @Module
 @InstallIn(SingletonComponent::class)
