+++ data/remote/src/main/java/com/jfmr/ac/test/data/remote/episode/datasource/RemoteEpisodesDataSource.kt
@@ -1,6 +1,6 @@
 package com.jfmr.ac.test.data.remote.episode.datasource
 
-import com.jfmr.ac.test.data.api.rickandmorty.episode.entity.EpisodeResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.episode.entity.EpisodeResponse
 import retrofit2.Response
 
 interface RemoteEpisodesDataSource {
