+++ data/remote/src/main/java/com/jfmr/ac/test/data/remote/episode/datasource/di/EpisodeDSModule.kt
@@ -1,17 +1,21 @@
-package com.jfmr.ac.test.data.di
+package com.jfmr.ac.test.data.remote.episode.datasource.di
 
-import com.jfmr.ac.test.data.remote.episode.datasource.RemoteEpisodesDataSource
 import com.jfmr.ac.test.data.remote.episode.RemoteEpisodesDataSourceImpl
-import com.jfmr.ac.test.data.remote.qualifier.QEpisodesDataSource
+import com.jfmr.ac.test.data.remote.episode.datasource.RemoteEpisodesDataSource
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import javax.inject.Qualifier
 import javax.inject.Singleton
+
+@Retention(AnnotationRetention.BINARY)
+@Qualifier
+annotation class QEpisodesDataSource
 
 @InstallIn(SingletonComponent::class)
 @Module
-interface DataSourceModule {
+interface EpisodeDSModule {
 
 
     @QEpisodesDataSource
