+++ data/remote/src/main/java/com/jfmr/ac/test/data/remote/extensions/NetworkExtensions.kt
