--- data/remote/src/main/java/com/jfmr/ac/test/data/remote/qualifier/Dispatchers.kt
@@ -1,16 +1,0 @@
-package com.jfmr.ac.test.data.remote.qualifier
-
-import javax.inject.Qualifier
-
-@Retention(AnnotationRetention.BINARY)
-@Qualifier
-annotation class DispatcherDefault
-
-@Retention(AnnotationRetention.BINARY)
-@Qualifier
-annotation class DispatcherIO
-
-@Retention(AnnotationRetention.BINARY)
-@Qualifier
-annotation class QEpisodesDataSource
-
