+++ data/remote/src/test/java/com/jfmr/ac/test/data/remote/episode/mapper/EpisodeExtensionsTest.kt
@@ -1,6 +1,6 @@
 package com.jfmr.ac.test.data.remote.episode.mapper
 
-import com.jfmr.ac.test.data.api.rickandmorty.episode.entity.EpisodeResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.episode.entity.EpisodeResponse
 import com.jfmr.ac.test.data.remote.episode.mapper.EpisodeExtensions.toDomain
 import com.jfmr.ac.test.domain.model.episode.Episode
 import com.jfmr.ac.test.tests.episodes.EpisodeUtils
