+++ data/repository/src/main/java/com/jfmr/ac/test/data/repository/episode/di/EpisodeRepositoryModule.kt
@@ -1,8 +1,6 @@
-package com.jfmr.ac.test.data.di
+package com.jfmr.ac.test.data.repository.episode.di
 
-import com.jfmr.ac.test.data.repository.character.CharacterRepositoryImpl
 import com.jfmr.ac.test.data.repository.episode.EpisodeRepositoryImpl
-import com.jfmr.ac.test.domain.repository.character.CharacterRepository
 import com.jfmr.ac.test.domain.repository.episode.EpisodeRepository
 import dagger.Binds
 import dagger.Module
@@ -13,20 +11,11 @@ import javax.inject.Singleton
 
 @Retention(AnnotationRetention.BINARY)
 @Qualifier
-annotation class QCharacterRepository
-
-@Retention(AnnotationRetention.BINARY)
-@Qualifier
 annotation class QEpisodesRepository
 
 @InstallIn(SingletonComponent::class)
 @Module
-interface RepositoryModule {
-
-    @QCharacterRepository
-    @Singleton
-    @Binds
-    fun bindCharactersRepository(implementation: CharacterRepositoryImpl): CharacterRepository
+interface EpisodeRepositoryModule {
 
     @QEpisodesRepository
     @Singleton
